% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFacets.R
\name{anova.TestFacets}
\alias{anova.TestFacets}
\title{anova TestFacets}
\usage{
\method{anova}{TestFacets}(object, type)
}
\arguments{
\item{object}{an object of class TestFacets}

\item{type}{which model to print}
}
\value{
summary of the lavaan results in the object
}
\description{
anova TestFacets
}
