% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BestFunctions.R
\name{bestDiscriminantItem}
\alias{bestDiscriminantItem}
\title{bestDiscriminantItem}
\usage{
bestDiscriminantItem(object, mainScale, discriminantScales)

bestDiscriminantItem(object, mainScale, discriminantScales)
}
\arguments{
\item{object}{Psychometric object}

\item{mainScale}{Scale to test which items are best}

\item{discriminantScales}{which scales to discriminate from}
}
\value{
sorted vector of items with highest discrimination first

sorted vector of items with highest discrimination first
}
\description{
bestDiscriminantItem

bestDiscriminantItem
}
