% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BestFunctions.R
\name{bestItems}
\alias{bestItems}
\title{Title}
\usage{
bestItems(object, scale, nItems)
}
\arguments{
\item{object}{Psychometric object}

\item{scale, }{scale to find best items for}

\item{nItems, }{number of items in new scale}
}
\value{
best item object
}
\description{
Title
}
