% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BestFunctions.R
\name{bestScales.Psychometric}
\alias{bestScales.Psychometric}
\title{BestScales}
\usage{
\method{bestScales}{Psychometric}(object, nItems, deleteItems = F)
}
\arguments{
\item{object}{Psychometric object}

\item{nItems}{Number of items in new scales}

\item{deleteItems}{F = allitems are left in the object. T: items not selected are deleted}
}
\value{
Psychometric object with scale means using the best nItems items
}
\description{
BestScales
}
