% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGFP.R
\name{corGFP.Psychometric}
\alias{corGFP.Psychometric}
\title{Get correlation GFPs}
\usage{
\method{corGFP}{Psychometric}(object)
}
\arguments{
\item{object}{a Psychometric object}
}
\value{
crrrelation matrix
}
\description{
Get correlation GFPs
}
