% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychometricHelper.R
\name{filter.Psychometric}
\alias{filter.Psychometric}
\title{filter}
\usage{
filter.Psychometric(object, keep)
}
\arguments{
\item{object}{a Psychometric object}

\item{keep}{a vector with boolean tne same lengths as the number of rows}
}
\value{
filtered object
}
\description{
filter
}
