% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychometricHelper.R
\name{filterV}
\alias{filterV}
\title{filterV}
\usage{
filterV(object, filterVector)

filterV(object, filterVector)
}
\arguments{
\item{object}{a Psychometric object to work with}

\item{filterVector}{a logical vector}
}
\value{
an updated Psychometric object

an updated Psychometric object
}
\description{
filterV

filterV
}
