% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychometricHelper.R
\name{getCommand.Psychometric}
\alias{getCommand.Psychometric}
\title{getCommand for Psychometric}
\usage{
\method{getCommand}{Psychometric}(object, ...)
}
\arguments{
\item{object}{a Psychometric object}

\item{...}{more commands}
}
\value{
text with commands to use
}
\description{
getCommand for Psychometric
}
