% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCommand.R
\name{getCommand}
\alias{getCommand}
\title{Get commands}
\usage{
getCommand(object, ...)
}
\arguments{
\item{object}{A reliability object}

\item{...}{scale Either "All" or a scale among those in ScaleNames}
}
\value{
a character string
}
\description{
Get commands
}
