% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychometricHelper.R
\name{getCommand.Reliability}
\alias{getCommand.Reliability}
\title{Get the reliability commands}
\usage{
\method{getCommand}{Reliability}(object, ...)
}
\arguments{
\item{object}{A reliability object}

\item{...}{scale Either "All" or a scale among those in ScaleNames}
}
\value{
a character string
}
\description{
Get the reliability commands
}
