% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychometricHelper.R
\name{getCommand.TestFacets}
\alias{getCommand.TestFacets}
\title{getCommand TestFacets}
\usage{
\method{getCommand}{TestFacets}(object, ...)
}
\arguments{
\item{object}{a TestFacets object created by TestFacets function}

\item{...}{model either "All" or a number}
}
\value{
text with commands to use
}
\description{
getCommand TestFacets
}
