% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGFP.R
\name{getCommand2.Psychometric}
\alias{getCommand2.Psychometric}
\title{Get commands for models}
\usage{
getCommand2.Psychometric(object, model)
}
\arguments{
\item{object}{a TestFacet object}

\item{model}{a model number}
}
\value{
a model
}
\description{
Get commands for models
}
