% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetThings.R
\name{getData.Psychometric}
\alias{getData.Psychometric}
\title{getData}
\usage{
\method{getData}{Psychometric}(object, scales = NULL, otherVar = T, scaleFrame = T, itemFrames = T)
}
\arguments{
\item{object}{the object to get data from}

\item{scales}{scales to extract, if null all scales}

\item{otherVar}{if true also other variables}

\item{scaleFrame}{if true also scales}

\item{itemFrames}{if true also items}
}
\value{
a dataframe with all scales and other variables from the Psychometric object
}
\description{
getData
}
