% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychometricHelper.R
\name{getData}
\alias{getData}
\title{Gets data from Psychometric object}
\usage{
getData(object)
}
\arguments{
\item{object}{the object to get data from}
}
\value{
a dataframe with all scales and other variables
}
\description{
Gets data from Psychometric object
}
\examples{
object <- GetPsychometric(persData, c("Achievement", "Dutifulness", "Orderly"),
 responseScale = list(c(0,4)), itemLength = 4)
data <- getData(object)
}
