% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetThings.R
\name{getItemText}
\alias{getItemText}
\title{getItemText}
\usage{
getItemText(object, scales = NULL, printName = F)

getItemText(object, scales = NULL, printName = F)
}
\arguments{
\item{object}{the object to get data from}

\item{scales}{scales to get item text, if null all scales}

\item{printName}{whether it should print the item texts}
}
\value{
a dataframe with all scales and other variables from the Psychometric object

a dataframe with all scales and other variables from the Psychometric object
}
\description{
getItemText

getItemText
}
