% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImputeMissing.R
\name{getMissing}
\alias{getMissing}
\title{Get number of missing}
\usage{
getMissing(object, items = F, scales = T, ...)
}
\arguments{
\item{object}{Psychometric object}

\item{items}{= F get number of missing on item level}

\item{scales}{T = do missing on scale level F = on item level}
}
\value{
a vector with names for scales and items, and number of missing in the cells
}
\description{
Makes it simple to do basic psychometrics
}
\details{
This is to chcek for missing to be sued buy other functions
}
