% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFacets.R
\name{getPredict}
\alias{getPredict}
\title{Get factor scores}
\usage{
getPredict(object, model)
}
\arguments{
\item{object}{a TestFacet object}

\item{model}{which model to return}
}
\value{
a dataframe with factor scores
}
\description{
Get factor scores
}
