% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReliabilityTest.R
\name{GetReliabilityTest}
\alias{GetReliabilityTest}
\title{Getting the Reliability object based on a Psychometric object}
\usage{
GetReliabilityTest(object, what = "Alpha", check = T, imp = "NULL", ...)
}
\arguments{
\item{object}{A Psychometric object}

\item{what}{what type of analyses: Alpha, Omega or Parallel}

\item{check}{whether to check keys in alpha}

\item{imp}{how to impute variables in alpha:}

\item{...}{more arguments}
}
\value{
A Reliability object based on a Psychometric object that can be used for analyses
}
\description{
Makes it simple to do basic psychometrics
}
