% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupInvariance.R
\name{GroupInvariance}
\alias{GroupInvariance}
\title{Group invariance test with CFA}
\usage{
GroupInvariance(
  object,
  scale,
  group,
  ordered = c(),
  fixed = F,
  fixedScales = F,
  parcel = F,
  fixedSubScales = c(),
  tries = 1,
  estimator = "ML",
  zeroVar = c(),
  delVar = c()
)
}
\arguments{
\item{object}{a Psychometric object}

\item{scale}{a Scale name}

\item{group}{a category variable included in "Other variables"}

\item{ordered}{whether the variables are ordinal or binary}

\item{fixed}{whether the bi-factor should be fixed to 1}

\item{fixedScales}{whether the scales should be fixed to 1}

\item{parcel}{whether the observed variables should be in parcels}

\item{fixedSubScales}{fix these subscales to 1}

\item{tries}{how many attempt to estimate model}

\item{estimator}{any estimator that is admissible for cfa}

\item{zeroVar}{var set to zero error variance}

\item{delVar}{delete some variables in a previous model}
}
\value{
TestFacet model
}
\description{
Group invariance test with CFA
}
