% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychometricHelper.R
\name{handleOutliers.Psychometric}
\alias{handleOutliers.Psychometric}
\title{Handle outliers methods}
\usage{
\method{handleOutliers}{Psychometric}(
  object,
  method = "Mahalanobis",
  limit = 0.001,
  missing = "None",
  otherVar = c()
)
}
\arguments{
\item{object}{a Psychometric object to work with}

\item{method}{there are four ways, "Mahalanobis", "Quartile", SD" and "Change"}

\item{limit}{the probability value for handeling an outlier}

\item{missing}{when "none", missing values are not handled, otherwise the method in missing will be used}

\item{otherVar}{whether the variables not among the scales shall be included, only for check and SD}
}
\value{
an updated Psychometric object
}
\description{
Handle outliers methods
}
\details{
Essentially this function takes a Psychometric object and checks for
outliers. The resulting object is sometimes changed, for example Mahalanobis and
SD and Quartile filters out the outliers and return an updated Psychometric object.
'Change' winsorizes the scales, e.g., changes the data values to be inside the
distribution.
The limit argument defines an outlier in terms of probabilty, either based on
chi2 value or SD.
The missing argument impute (or deletes cases). See the ImputeMissing function.
Using the 'OtherVariables' flag makes it possible to also handle other variables
not included in some of the scales.
}
