% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychometricHelper.R
\name{handleOutliers}
\alias{handleOutliers}
\title{Handle outliers methods}
\usage{
handleOutliers(
  object,
  method = "Mahalanobis",
  limit = 0.001,
  missing = "None",
  otherVar = F
)
}
\value{
an updated Psychometric object
}
\description{
Handle outliers methods
}
