% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImputeMissing.R
\name{imputeMissing}
\alias{imputeMissing}
\title{Summary of imputeMissing}
\usage{
imputeMissing(object, handleMissing = "Listwise", scales = T, ...)
}
\arguments{
\item{object}{A Reliability object}

\item{handleMissing}{can be:  Listwise, Mean, Impute, Bayesian, Regression, Pmm, BayesianMean, and check}

\item{scales}{T = do missing on scale level F = on item level}

\item{...}{k = can be used to create BayesianMean based on k imputations
printFlag = T can be used to get information about imputations}
}
\value{
A Psychometric object that can be used for analyses
}
\description{
Makes it simple to do basic psychometrics
}
