% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeParcels.R
\name{makeParcels}
\alias{makeParcels}
\title{Title}
\usage{
makeParcels(object, scales, nItems, random = FALSE)
}
\arguments{
\item{object}{Psychometric object}

\item{scales, }{scale to find best items for}

\item{nItems, }{number of items in new scale}

\item{random, }{random parcels}
}
\value{
best item object
}
\description{
Title
}
