% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeParcels.R
\name{makeThreeParcels}
\alias{makeThreeParcels}
\title{makeThreeParcels}
\usage{
makeThreeParcels(object, scales, random = FALSE)
}
\arguments{
\item{object}{Psychometric object}

\item{scales, }{scale to find best items for}

\item{random, }{random parcels}
}
\value{
best item object
}
\description{
makeThreeParcels
}
