% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFacets.R
\name{myAnova.Psychometric}
\alias{myAnova.Psychometric}
\title{Title}
\usage{
\method{myAnova}{Psychometric}(object, type)
}
\arguments{
\item{object}{an object of class TestFacets}

\item{mdoel}{which model to print}
}
\value{
summary of the lavaan results in the object
}
\description{
Title
}
