% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFacets.R
\name{myAnova}
\alias{myAnova}
\title{myAnova}
\usage{
myAnova(object, type = "Bifactor")
}
\arguments{
\item{object}{an object of class TestFacets}

\item{type}{an be "Bifactor" or "Hiearchical" or "B" or "H"}
}
\description{
myAnova
}
