% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFacets.R
\name{mySummary.Psychometric}
\alias{mySummary.Psychometric}
\title{Title}
\usage{
\method{mySummary}{Psychometric}(object, model = NULL, standardized = F)
}
\arguments{
\item{object}{an object of class TestFacets}

\item{model}{which model to print}

\item{standardized}{whether the standardized values should be printed}
}
\value{
summary of the lavaan results in the object
}
\description{
Title
}
