% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFacets.R
\name{mySummary}
\alias{mySummary}
\title{mySummary to get all the output}
\usage{
mySummary(object, model = NULL, standardized = F)
}
\arguments{
\item{object}{an object of class TestFacets}

\item{model}{which model to print}

\item{standardized}{whether the standardized values should be printed}
}
\value{
Results
}
\description{
mySummary to get all the output
}
