% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychometricHelper.R
\name{names.Psychometric}
\alias{names.Psychometric}
\title{names}
\usage{
\method{names}{Psychometric}(x)
}
\arguments{
\item{x}{a Psychometric object to work with}
}
\value{
the names of Scale and Other variables
}
\description{
names
}
\examples{
dataObject <- GetPsychometric(persData,
scaleNames = c("Extraversion", "Agreeableness",
               "Conscientiousness", "Neuroticism",
               "Openness"),
responseScale = list(c(0,36)),
itemLength = 4)
names(dataObject)
}
