% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetThings.R
\name{namesP.Psychometric}
\alias{namesP.Psychometric}
\title{namesP}
\usage{
\method{namesP}{Psychometric}(x)
}
\arguments{
\item{x}{a Psychometric object to work with}
}
\value{
the names of Scale and Other variables
}
\description{
namesP
}
