% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetThings.R
\name{namesP}
\alias{namesP}
\title{namesP}
\usage{
namesP(x)
}
\arguments{
\item{x}{a Psychometric object to work with}
}
\value{
the names of Scale and Other variables
}
\description{
namesP
}
\examples{
dataObject <- GetPsychometric(persData,
scaleNames = c("Extraversion", "Agreeableness",
               "Conscientiousness", "Neuroticism",
               "Openness"),
responseScale = list(c(0,36)),
itemLength = 4)
namesP(dataObject)
}
