% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{persData}
\alias{persData}
\title{Personality inventory NB5I with 20 subscales}
\format{
A data frame with 200 rows and 125 variables:
\describe{
\item{ID}{identity, anonymous}
\item{Sex}{forGender, anonymous}
\item{ExtrHappy}{Item for extraversion scale}
\item{NeuroVulnerable}{Item for neuroticism scale}
\item{ConscDiligent}{Item for conscientousness scale}
\item{OpenIntellectual}{Item for openness scale}
\item{AgreeCooperative}{Item for agreeableness scale}
\item{NeuroDepressed}{Item for neuroticism scale}
\item{ExtrFriendly}{Item for the extraversion scale}
\item{OpenArtistic}{Item for the openness scale}
\item{ExtrSensationSeeking}{Item for the extraversion scale}
\item{ConscAchivement}{Item for the conscientousness scale}
\item{ConscOrderly}{Item for the conscientousness scale}
\item{NeuroAnxioius}{Item for the neuroticism scale}
\item{NeuroIrritaböle}{Item for the neuroticism scale}
\item{OpenInnerLife}{Item for the openness scale}
\item{AgreeAltruistic}{Item for the agreeableness scale}
\item{AgreeMoral}{Item for the agreeableness scale}
\item{ExtrActive}{Item for the extraversion scale}
\item{OpenVariabilitySeeking}{Item for the openness scale}
\item{AgreeSympathetic}{Item for the agreeableness scale}
\item{ConscEffective}{Item for the conscientousness scale}
...
}
}
\source{
None
}
\usage{
persData
}
\description{
A dataset with  20 scales and 5 factors (BigFive)
}
\keyword{datasets}
