% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot.Psychometric}
\alias{plot.Psychometric}
\title{Plot scales from Psychometric}
\usage{
\method{plot}{Psychometric}(x, y, ...)
}
\arguments{
\item{x}{a Psychometric object}

\item{y}{optional not used}

\item{...}{extra argument is: scale = scaleName}
}
\value{
nothing
}
\description{
Plot scales from Psychometric
}
