% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReliabilityTest.R
\name{print.Reliability}
\alias{print.Reliability}
\title{Prints the results of the Reliabilty object}
\usage{
\method{print}{Reliability}(x, ...)
}
\arguments{
\item{x}{A Reliability object}

\item{...}{Other arguments}
}
\value{
A Reliability object that can be used for analyses
}
\description{
Makes it simple to do basic psychometrics
}
