% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFacets.R
\name{print.TestFacets}
\alias{print.TestFacets}
\title{Title}
\usage{
\method{print}{TestFacets}(object, model = NULL, standardized = F)
}
\arguments{
\item{object}{an object of class TestFacets}

\item{model}{which model to print}

\item{standardized}{whether standardized solution should be printed}
}
\value{
summary of the lavaan results in the object
}
\description{
Title
}
