% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.R
\name{pSummary.Psychometric}
\alias{pSummary.Psychometric}
\title{pSummary}
\usage{
pSummary.Psychometric(
  object,
  mean = T,
  sd = T,
  SE = T,
  skew = T,
  kurtosis = T,
  min = T,
  max = T,
  omega = T,
  n = T,
  plots = F
)

pSummary.Psychometric(
  object,
  mean = T,
  sd = T,
  SE = T,
  skew = T,
  kurtosis = T,
  min = T,
  max = T,
  omega = T,
  n = T,
  plots = F
)
}
\arguments{
\item{object}{The Psychometric object}

\item{mean}{T: Shown, F: not shown}

\item{sd}{T: Shown, F: not shown}

\item{SE}{T: Shown, F: not shown}

\item{skew}{T: Shown, F: not shown}

\item{kurtosis}{T: Shown, F: not shown}

\item{min}{T: Shown, F: not shown}

\item{max}{T: Shown, F: not shown}

\item{omega}{T: Shown, F: not shown}

\item{n}{T: Shown, F: not shown}

\item{plots}{F: no plots are shown, T: plots from psych::omega are shown}
}
\value{
the summary i a list object

the summary i a list object
}
\description{
Includes a set of useful estimates for the scales

Includes a set of useful estimates for the scales
}
\examples{
object <- GetPsychometric(persData, c("Achievement", "Dutifulness", "Orderly"),
 responseScale = list(c(0,4)), itemLength = 4)
pSsummary(object)
object <- GetPsychometric(persData, c("Achievement", "Dutifulness", "Orderly"),
 responseScale = list(c(0,4)), itemLength = 4)
pSsummary(object)
}
