% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFacets.R
\name{RunFA}
\alias{RunFA}
\title{A Psychometric FA estimator}
\usage{
RunFA(object, subscales, ...)

RunFA(object, subscales, ...)
}
\arguments{
\item{object}{an of class TestFacets}

\item{subscales}{the subscales that are included}

\item{...}{extra argument to the fa}
}
\value{
a psych::fa result based on the selected subscales

a psych::fa result based on the selected subscales
}
\description{
A Psychometric FA estimator

A Psychometric FA estimator
}
