% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFacets.R
\name{RunPCA}
\alias{RunPCA}
\title{A Psychometric FA estimator}
\usage{
RunPCA(object, subscales, ...)
}
\arguments{
\item{object}{an of class TestFacets}

\item{subscales}{the subscales that are included}

\item{...}{extra argument to the fa}
}
\value{
a psych::fa result based on the selected subscales
}
\description{
A Psychometric FA estimator
}
