% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetThings.R
\name{selectP.Psychometric}
\alias{selectP.Psychometric}
\title{select}
\usage{
\method{selectP}{Psychometric}(object, scales)
}
\arguments{
\item{object}{a Psychometric object}

\item{scales}{the scales that should be included in the result}
}
\value{
an Psychometric object with only the scales
}
\description{
select
}
