% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Split.R
\name{split.Psychometric}
\alias{split.Psychometric}
\title{Split Psychometric}
\usage{
\method{split}{Psychometric}(object, group, f, ...)
}
\arguments{
\item{object}{A Psychometric object}

\item{group}{A group variables among other variables}

\item{f}{A function to apply to the groups}

\item{...}{more arguments to the f function}
}
\value{
all results from the f function
}
\description{
Makes it simple to work with groups
}
\examples{
object <- GetPsychometric(persData, c("Achievement", "Dutifulness", "Orderly"),
 splitPsychometric(object, group = "Sex", f = Summary)

}
