% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Split.R
\name{splitP}
\alias{splitP}
\title{Split Psychometric}
\usage{
splitP(object, f = NULL, group, ...)
}
\arguments{
\item{object}{A Psychometric object}

\item{f}{A function to apply to the groups}

\item{group}{A group variables among other variables}

\item{...}{more arguments to the f function}
}
\value{
all results from the f function
}
\description{
Makes it simple to work with groups
}
