% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychometricHelper.R
\name{splitPsychometric}
\alias{splitPsychometric}
\title{Split Psychometric}
\usage{
splitPsychometric(object, group, f, ...)
}
\arguments{
\item{object}{A Psychometric object}

\item{group}{A group variables among other variables}

\item{f}{A function to apply to the groups}

\item{...}{more arguments to the f function}
}
\value{
all results from the f function
}
\description{
Makes it simple to work with groups
}
