% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetThings.R
\name{[.Psychometric}
\alias{[.Psychometric}
\title{select'[]'}
\usage{
\method{[}{Psychometric}(x, i)
}
\arguments{
\item{x}{a Psychometric object}

\item{i}{the index}
}
\value{
an Psychometric object only the indexed rows
}
\description{
select'[]'
}
