% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPsychometric.R, R/Summary.R
\name{summary.Psychometric}
\alias{summary.Psychometric}
\title{summary for Psychometric}
\usage{
\method{summary}{Psychometric}(
  object,
  mean = T,
  sd = T,
  SE = T,
  skew = T,
  kurtosis = T,
  min = T,
  max = T,
  omega = T,
  n = T,
  plots = F
)

\method{summary}{Psychometric}(
  object,
  mean = T,
  sd = T,
  SE = T,
  skew = T,
  kurtosis = T,
  min = T,
  max = T,
  omega = T,
  n = T,
  plots = F
)
}
\arguments{
\item{object}{The Psychometric object}

\item{mean}{T: Shown, F: not shown}

\item{sd}{T: Shown, F: not shown}

\item{SE}{T: Shown, F: not shown}

\item{skew}{T: Shown, F: not shown}

\item{kurtosis}{T: Shown, F: not shown}

\item{min}{T: Shown, F: not shown}

\item{max}{T: Shown, F: not shown}

\item{omega}{T: Shown, F: not shown}

\item{n}{T: Shown, F: not shown}

\item{plots}{F: no plots are shown, T: plots from psych::omega are shown}

\item{...}{There are several extra arguments to delete from the output
mean: should the mean be displayed T/F
sd:  should the sd be displayed T/F
SE:  should the SE be displayed T/F
skew:  should the skew be displayed T/F
kurtosis:  should the kurtosis be displayed T/F
min:  should the min be displayed T/F
max:  should the max be displayed T/F
alpha:  should the alpha be displayed T/F. Warnings at Level
n:  should the n be displayed T/F
plots:  should the plots be displayed T/F}
}
\value{
Nothing

the summary i a list object
}
\description{
Includes a set of useful estimates for the scales
}
\examples{
object <- GetPsychometric(persData, c("Achievement", "Dutifulness", "Orderly"),
   responseScale = list(c(0,4)), itemLength = 4)
summary(object)
object <- GetPsychometric(persData, c("Achievement", "Dutifulness", "Orderly"),
 responseScale = list(c(0,4)), itemLength = 4)
summary(object)
}
