% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReliabilityTest.R
\name{summary.Reliability}
\alias{summary.Reliability}
\title{Summary of Reliability}
\usage{
\method{summary}{Reliability}(object, ...)
}
\arguments{
\item{object}{A Reliability object}

\item{...}{which scale}
}
\value{
A Reliability object that can be used for analyses
}
\description{
Makes it simple to do basic psychometrics
}
