% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFacets.R
\name{summary.TestFacets}
\alias{summary.TestFacets}
\title{summary for TestFacets}
\usage{
\method{summary}{TestFacets}(object, model = NULL, standardized = F)
}
\arguments{
\item{object}{an object of class TestFacets}

\item{model}{which model to print}

\item{standardized}{whether the standardized values should be printed}
}
\value{
summary of the lavaan results in the object
}
\description{
summary for TestFacets
}
