% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFactors.R
\name{TestCongruence}
\alias{TestCongruence}
\title{TestCongruence}
\usage{
TestCongruence(
  objectList,
  fms = "ml",
  rotates = "oblimin",
  itemNums = 50,
  factors = 4,
  printFlag = F,
  printLoadings = F,
  printAll = F
)

TestCongruence(
  objectList,
  fms = "ml",
  rotates = "oblimin",
  itemNums = 50,
  factors = 4,
  printFlag = F,
  printLoadings = F,
  printAll = F
)
}
\arguments{
\item{objectList}{Psychometric objects to compare}

\item{fms}{Which method from psych::fa to extract factors}

\item{rotates}{Vector with psych::fa rotation names to be compared}

\item{itemNums}{Vector of integers used to compare scales based on fewer items. Items selected with psych::BestScales}

\item{factors}{Vector of integer of number of factors to test in each model}

\item{printFlag}{Bool for whether the congruence tables should be printed}

\item{printLoadings}{Bool for whether the loadings should be printed}

\item{printAll}{Bool for whether the factor model should be printed}
}
\value{
A list with all congruent tested models with their respective congruence value (mean of best match)

A list with all congruent tested models
}
\description{
TestCongruence

TestCongruence
}
