% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFacets.R
\name{TestFacets.Psychometric}
\alias{TestFacets.Psychometric}
\title{Facet test with CFA}
\usage{
\method{TestFacets}{Psychometric}(
  object,
  scales,
  subscales,
  fixed = F,
  fixedScales = F,
  parcel = F,
  fixedSubScales = c(),
  tries = 1,
  estimator = "ML",
  zeroVar = c()
)
}
\arguments{
\item{object}{a Psychometric object}

\item{subscales}{a vector of facet names}

\item{fixed}{whether the bi-factor should be fixed to 1}

\item{fixedScales}{whether the scales should be fixed to 1}

\item{parcel}{whether the observed variables should be in parcels}

\item{fixedSubScales}{fix these subscales to 1}

\item{estimator}{any estimator that is admissible for cfa}

\item{scale}{a Scale name}
}
\value{
a TestFacet model
}
\description{
Facet test with CFA
}
