% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFacets.R
\name{TestFacets}
\alias{TestFacets}
\title{Title}
\usage{
TestFacets(object, scale, subscales)
}
\arguments{
\item{object}{a Psychometric object}

\item{scale}{a Scale name}

\item{subscales}{a vector of facet names}
}
\value{
results
}
\description{
Title
}
\examples{
object <- GetPsychometric(persData, c("Achievement", "Dutifulness", "Orderly"),
   responseScale = list(c(0,4)), itemLength = 4)
TestFacets(object, "Cons", c("Achievement", "Dutifulness", "Orderly"))
}
