% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeParcels.R
\name{testRandomParcels}
\alias{testRandomParcels}
\title{testRandomParcels}
\usage{
testRandomParcels(object, model, scales, k = 31, random = FALSE)
}
\arguments{
\item{object}{Psychometric object}

\item{model}{Should be the model that test for}

\item{scales, }{scale to find best items for}

\item{k, }{number of items in new scale}

\item{random, }{random parcels}
}
\value{
best item object
}
\description{
testRandomParcels
}
