% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychometricHelper.R
\name{write.csv.Psychometric}
\alias{write.csv.Psychometric}
\title{Write all data}
\usage{
write.csv.Psychometric(object, fileName, colnames = T, rownames = F)
}
\arguments{
\item{object}{A psychometric object}

\item{fileName}{A filename with pathinformation}

\item{col.names}{T if you like to write column names}

\item{row.names}{T if you like to write row names}
}
\description{
Write all data
}
