% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychometricHelper.R
\name{writeP.Psychometric}
\alias{writeP.Psychometric}
\title{Write all data}
\usage{
\method{writeP}{Psychometric}(object, File, colnames = T, rownames = F)
}
\arguments{
\item{object}{A psychometric object}

\item{File}{A filename with pathinformation}

\item{colnames}{T if you like to write column names}

\item{rownames}{T if you like to write row names}
}
\description{
Write all data
}
