% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFacets.R
\name{anova.TestFacets}
\alias{anova.TestFacets}
\title{anova TestFacets}
\usage{
anova.TestFacets(object, ...)
}
\arguments{
\item{object}{an object of class TestFacets}

\item{...}{type which model to print and more extra}
}
\value{
summary of the lavaan results in the object
}
\description{
anova TestFacets
}
