% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BestFunctions.R
\name{bestScales}
\alias{bestScales}
\title{BestScales}
\usage{
bestScales(object, nItems, deleteItems = F)
}
\arguments{
\item{object}{Psychometric object}

\item{nItems}{Number of items in new scales}

\item{deleteItems}{F = allitems are left in the object. T: items not selected are deleted}
}
\value{
Psychometric object with scale means using the best nItems items
}
\description{
BestScales
}
\examples{
object <- GetPsychometric(persData, c("Achievement", "Dutifulness", "Orderly"),
 responseScale = list(c(0,4)), itemLength = 4)
res <- bestScales(object,  5)
}
