% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychometricHelper.R
\name{filterP.Psychometric}
\alias{filterP.Psychometric}
\title{filter}
\usage{
filterP.Psychometric(object, ...)

filterP.Psychometric(object, ...)
}
\arguments{
\item{object}{a Psychometric object}

\item{keep}{a vector with boolean tne same lengths as the number of rows}
}
\value{
filtered object

filtered object
}
\description{
filter

filter
}
