% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReliabilityTest.R
\name{getCommand}
\alias{getCommand}
\title{Get the reliability commands}
\usage{
getCommand(object, scale)
}
\arguments{
\item{object}{A reliability object}

\item{scale}{Either "All" or a scale among those in ScaleNames}
}
\value{
a character string
}
\description{
Get the reliability commands
}
