% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGFP.R
\name{getGFP.Psychometric}
\alias{getGFP.Psychometric}
\title{Get All GFPs}
\usage{
\method{getGFP}{Psychometric}(
  object,
  scales = NULL,
  parceling = F,
  fixed = F,
  fixedScales = F,
  fixedSubScales = c()
)
}
\arguments{
\item{object}{a TestFacet object}

\item{scales}{should scales be fixed to 1}

\item{parceling}{should facets be parceled}

\item{fixed}{should general factors be fixed to 1}
}
\value{
an extended TestFacet object
}
\description{
Get All GFPs
}
