% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPsychometric.R
\name{GetPsychometric}
\alias{GetPsychometric}
\title{Psychometric analyses of Likert scales}
\usage{
GetPsychometric(
  data,
  scaleNames,
  responseScale = list(c(1, 5)),
  typeSum = "Mean",
  itemLength = 6,
  reverse = T,
  idVar = "ID",
  name = "Psychometric",
  itemDictionary = NULL
)
}
\arguments{
\item{data}{datafile with items columns from one or more scales}

\item{scaleNames}{description}

\item{responseScale}{description}

\item{typeSum}{description}

\item{itemLength}{description}

\item{reverse}{description}

\item{idVar}{description}

\item{name}{description}

\item{itemDictionary}{description}
}
\description{
Psychometric analyses of Likert scales
}
\examples{
dat <- as.data.frame(list(pItem1 = c(2,3,4,4,3,4,3,4), pItem2 = c(2,3,4,4,2,4,2,3)))
GetPsychometric(dat, "p", responseScale = list(c(0,4)), itemLength = 1)

}
