% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPsychometric.R
\name{GetPsychometric}
\alias{GetPsychometric}
\title{Getting the Psychometric class}
\usage{
GetPsychometric(
  data,
  scaleNames,
  responseScale = list(c(1, 5)),
  typeSum = "Mean",
  itemLength = 6,
  reverse = T,
  idVar = "ID",
  name = "Psychometric",
  itemDictionary = NULL,
  itemList = NULL,
  reverseList = c(),
  missings = NULL
)
}
\arguments{
\item{data}{A dataframe with the items and eventual extra variables}

\item{scaleNames}{A vector with the names of the scales}

\item{responseScale}{A list with vectors of response intervalls, either one (if all are the same) or one for each scale}

\item{typeSum}{A functions used for summarizing the items to scales}

\item{itemLength}{The number of significant characters in items}

\item{reverse}{A boolean that sets whether items ending with 'R' should be reversed}

\item{idVar}{The name of the case identifier (if none a case number variable called ID will be created)}

\item{name}{The name of the object (optional)}

\item{itemDictionary}{a textfile with information about all scales, items and whether they should be reversed}

\item{itemList}{items do not conform to above they can be addressed by number instead, new item names will be created based on scale names}

\item{reverseList}{a vector with numbers for all items that should be reversed, if any}

\item{missings}{a vector with numbers to be converted to NA, e.g., -99 should be converted to NA}
}
\value{
A Psychometric object that can be used for analyses.
}
\description{
Makes it simple to do basic psychometrics of scales
}
\details{
The GetPsychometric functions create a class that can be used to simplify
psychometric evaluation of a scale or an inventory. The function takes a dataframe
with items that are automatically aggregated to scales based on one of two different
methods.
\enumerate{
\item Either you name the item columns in a consistent way based on the scale names
you like to have, e.g., all items from the extraversion scale starts with 'extr'
\item You send an list with vectors that have the column numbers of the items for
each scale
}

You always always provide a vector of scale names, e.g., c("Achievement", "Dutifulness", "Orderly"),
to the argument scaleNames and you also provide a itemLength argument, that define the number of characters
that define you items, e.g., itemLength = 4 suggesting that the first four characters
defines the items belonging to the scale. If an item should be reversed you should
add an 'R' to the end of the name and also set the reverse flag to TRUE
When you are Using alternative 1 above and has Achievement as scale name, the items should
start with column name "Achi", like "Achive10" or "Achive12".
Using alternative 2 above, you add an itemList with column numbers like this:
list(c(4,5,6,7), c(8,9,10,11), c(12,13,14,15)) representing three scales. Item than
need to be reversed should be defined in a vector and sent to reverseList
If there is an identity variable you can define it with the idVar argument.
If there are missing values use define a vector with column numbers
and send it with the missings argument.

The object include a number of useful dataframes:
\enumerate{
\item A dataframe called 'ScaleFrame' that include the value of the scale
\item A list of dataframes calles 'ScaleItemFrames' that include the items included
in each of the scales.
\item All scale names in the 'ScaleNames' vector
\item Original data are in the 'OriginalData' frame
\item The variables not included in the scales are in the "OtherVariables" dataframe
together with the ID variable
In addition there are other objects that are used together with some of the
methods/functions that can be called with the object
}
}
\examples{
dataObject <- GetPsychometric(persData,
scaleNames = c("Extraversion", "Agreeableness",
               "Conscientiousness", "Neuroticism",
               "Openness"),
responseScale = list(c(0,24)),
itemLength = 4)

dataObject <- GetPsychometric(persData,
                scaleNames = c("Extraversion", "Agreeableness",
                              "Conscientiousness", "Neuroticism",
                             "Openness"),
                responseScale = list(c(0,24)),
                itemList = list(c(2,9,11,19), c(7,17,18, 21),
                c(5, 12,13), c(4,8,14,15), c(6,10, 16,20)),
                itemLength = 4)

}
