% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPsychometric.R
\name{GetPsychometric}
\alias{GetPsychometric}
\title{Getting the Psychometric class}
\usage{
GetPsychometric(
  data,
  scaleNames,
  responseScale = list(c(1, 5)),
  typeSum = "Mean",
  itemLength = 6,
  reverse = T,
  idVar = "ID",
  name = "Psychometric",
  itemDictionary = NULL,
  itemList = NULL,
  reverseList = c(),
  missings = NULL
)
}
\arguments{
\item{data}{A dataframe with the items and eventual extra variables}

\item{scaleNames}{A vector with the names of the scales}

\item{responseScale}{A list with vectors of response intervalls, either one (if all are the same) or one for each scale}

\item{typeSum}{A functions used for summarizing the items to scales}

\item{itemLength}{The number of significant characters in items}

\item{reverse}{A boolean that sets whether items ending with 'R' should be reversed}

\item{idVar}{The name of the case identifier (if none a case number variable called ID will be created)}

\item{name}{The name of the object (optional)}

\item{itemDictionary}{a textfile with information about scale, items and sign}

\item{itemList}{a textfile with information about scale, items and sign}

\item{reverseList}{a textfile with information about scale, items and sign}

\item{missings}{a vector with numbers to be converted to NA}
}
\value{
A Psychometric object that can be used for analyses
}
\description{
Makes it simple to do basic psychometrics of scales
}
\examples{
object <- GetPsychometric(persData, c("Achievement", "Dutifulness", "Orderly"),
 responseScale = list(c(0,4)), itemLength = 4)

}
