% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReliabilityTest.R
\name{GetReliabilityTest}
\alias{GetReliabilityTest}
\title{Getting the Reliability object based on a Psychometric object}
\usage{
GetReliabilityTest(object, what = "Alpha", ...)
}
\arguments{
\item{object}{A Psychometric object}

\item{what}{what type of analyses: Alpha, Omega or Parallel}

\item{...}{gh}
}
\value{
A Reliability object that can be used for analyses
}
\description{
Makes it simple to do basic psychometrics
}
\examples{
dat <- as.data.frame(list(pItem1 = c(2,3,4,2,3,4,3,4), pItem2 = c(2,3,5,4,2,0,4,3)))
object <- GetPsychometric(dat, "p", responseScale = list(c(0,4)), itemLength = 1)
relObject <- GetReliabilityTest(object)
}
