% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReliabilityTest.R
\name{GetReliabilityTest}
\alias{GetReliabilityTest}
\title{Getting the Reliability object based on a Psychometric object}
\usage{
GetReliabilityTest(object, what = "Alpha", ...)
}
\arguments{
\item{object}{A Psychometric object}

\item{what}{what type of analyses: Alpha, Omega or Parallel}

\item{...}{gh}
}
\value{
A Reliability object that can be used for analyses
}
\description{
Makes it simple to do basic psychometrics
}
\examples{
object <- GetPsychometric(persData, c("Achievement", "Dutifulness", "Orderly"),
 responseScale = list(c(0,4)), itemLength = 4)
relObject <- GetReliabilityTest(object)
}
