% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetThings.R, R/PsychometricHelper.R
\name{getSubScaleNames}
\alias{getSubScaleNames}
\title{getSubScaleNames}
\usage{
getSubScaleNames(object)

getSubScaleNames(object)

getSubScaleNames(object)

getSubScaleNames(object)
}
\arguments{
\item{object}{a Psychometric object}

\item{scales}{if NULL all scales otherswise a subset of scales in a Psychometric object}
}
\value{
a list of all names starting with the

a list of all names starting with the

a list of all names starting with the

a list of all names starting with the
}
\description{
getSubScaleNames

getSubScaleNames

getSubScaleNames

getSubScaleNames
}
\details{
When the ScaleItemFrames is created it changes the variable names of
the items. This functions return all the names together with the scale names.
The functions can be used to check that everything has been defined correctly.

When the ScaleItemFrames is created it changes the variable names of
the items. This functions return all the names together with the scale names.
The functions can be used to check that everything has been defined correctly.

When the ScaleItemFrames is created it changes the variable names of
the items. This functions return all the names together with the scale names.
The functions can be used to check that everything has been defined correctly.

When the ScaleItemFrames is created it changes the variable names of
the items. This functions return all the names together with the scale names.
The functions can be used to check that everything has been defined correctly.
}
