% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychometricHelper.R
\name{getSubScaleNames}
\alias{getSubScaleNames}
\title{Title}
\usage{
getSubScaleNames(object)
}
\arguments{
\item{object}{a Psychometric object}
}
\value{
names of all scale and items
}
\description{
Title
}
