% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychometricHelper.R
\name{handleOutliers.Psychometric}
\alias{handleOutliers.Psychometric}
\title{Handle outliers methods}
\usage{
\method{handleOutliers}{Psychometric}(
  object,
  method = "Mahalanobis",
  limit = 0.001,
  missing = "None",
  otherVar = c()
)
}
\arguments{
\item{object}{a Psychometric object to work with}

\item{method}{there are three ways, "Mahalanobis", "Quartil", SD" and "Change"}

\item{limit}{the probability value for handeling an outlier}

\item{missing}{when "none", missing values are not handled, otherwise the method in missing will be used}

\item{otherVar}{whether the variables not among the scales shall be included, only for check and SD}
}
\value{
an updated Psychometric object
}
\description{
Handle outliers methods
}
