% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychometricHelper.R
\name{handleOutliers}
\alias{handleOutliers}
\title{Handle outliers methods}
\usage{
handleOutliers(object, method = "Mahalanobis", limit = 0.001, missing = "None")
}
\arguments{
\item{object}{a Psychometric object to work with}

\item{method}{there are three ways, "Mahalanobis", "SD" and "Change"}

\item{limit}{the probability value for handeling an outlier}

\item{missing}{when "none", missing values are not handled, otherwise the method in missing will be used}
}
\value{
a Psychometric object with handled outliers
}
\description{
Handle outliers methods
}
\examples{
object <- GetPsychometric(persData, c("Achievement", "Dutifulness", "Orderly"),
 responseScale = list(c(0,4)), itemLength = 4)
newObject <- handleOutliers(object)
}
