% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImputeMissing.R
\name{imputeMissing}
\alias{imputeMissing}
\title{Summary of imputeMissing}
\usage{
imputeMissing(object, handleMissing = "Listwise", scales = T, ...)
}
\arguments{
\item{object}{A Reliability object}

\item{handleMissing}{A Reliability object}

\item{scales}{do missing on scale level}

\item{...}{commands to missing}
}
\value{
A Psychometric object that can be used for analyses
}
\description{
Makes it simple to do basic psychometrics
}
\examples{
dat <- as.data.frame(list(pItem1 = c(2,3,4,4,3,4,NA,4), pItem2 = c(2,3,4,4,2,4,2,3)))
myObject <- GetPsychometric(dat, "p", responseScale = list(c(0,4)), itemLength = 1)
myObject <- imputeMissing(myObject)
}
