% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychometricHelper.R
\name{names.Psychometric}
\alias{names.Psychometric}
\title{Names}
\usage{
\method{names}{Psychometric}(x)
}
\arguments{
\item{x}{a Psychometric object to work with}
}
\value{
the names of Scale and Other variables

examples
object <- GetPsychometric(persData, c("Achievement", "Dutifulness", "Orderly"),
names(object)
}
\description{
Names
}
