% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{persData}
\alias{persData}
\title{Personality inventory NB5I with 120 items}
\format{
A data frame with 200 rows and 125 variables:
\describe{
\item{ID}{identity, anonymous}
\item{AchieveACP01y}{Item for achievement scale}
\item{AchieveACP03y}{Item for achievement scale}
\item{AchieveACP1r}{Item for achievement scale}
\item{AchieveACP3p}{Item for achievement scale}
\item{AchieveACPR7c}{Item for achievement scale}
\item{AchieveACPR8x}{Item for achievement scale}
\item{ActiveAEA02x}{Item for the active scale}
\item{ActiveAEA02y}{Item for the active scale}
\item{ActiveAEA05y}{Item for the active scale}
\item{ActiveAEA08z}{Item for the active scale}
\item{ActiveAEA13s}{Item for the active scale}
\item{ActiveAEA17x}{Item for the active scale}
\item{AltruisAAA01z}{Item for the altruistic scale}
\item{AltruisAAA02y}{Item for the altruistic scale}
\item{AltruisAAA03y}{Item for the altruistic scale}
\item{AltruisAAA04y}{Item for the altruistic scale}
\item{AltruisAAA05z}{Item for the altruistic scale}
\item{AltruisAAA06v}{Item for the altruistic scale}
\item{ArtistiAOA01y}{Item for the artistic scale}
\item{ArtistiAOA02y}{Item for the artistic scale}
\item{ArtistiAOA03y}{Item for the artistic scale}
\item{ArtistiAOA04y}{Item for the artistic scale}
\item{ArtistiAOA6x}{Item for the artistic scale}
\item{ArtistiAOA7x}{Item for the artistic scale}
\item{CheckCHECK1E}{Item for check scale}
\item{CheckCHECK2E}{Item for check scale}
\item{CheckCHECK3E}{Item for check scale}
\item{CheckCHECK4E}{Item for check scale}
\item{CooperaAAC02x}{Item for the cooperative scale}
\item{CooperaAAC05v}{Item for the cooperative scale}
\item{CooperaAAC2p}{Item for the cooperative scale}
\item{CooperaAAC3s}{Item for the cooperative scale}
\item{CooperaAAC4s}{Item for the cooperative scale}
\item{CooperaAAC6t}{Item for the cooperative scale}
\item{DutifulACD01y}{Item for dutiful scale}
\item{DutifulACD03y}{Item for dutiful scale}
\item{DutifulACD05y}{Item for dutiful scale}
\item{DutifulACD06y}{Item for dutiful scale}
\item{DutifulACD2x}{Item for dutiful scale}
\item{DutifulACP05z}{Item for dutiful scale}
\item{ExitemeAEE01y}{Item for the exitement seeking scale}
\item{ExitemeAEE02y}{Item for the exitement seeking scale}
\item{ExitemeAES4t}{Item for the exitement seeking scale}
\item{ExitemeAESY02x}{Item for the exitement seeking scale}
\item{ExitemeAESY3x}{Item for the exitement seeking scale}
\item{ExitemeANES4x}{Item for the exitement seeking scale}
\item{FriendlAEF05y}{Item for the friendliness scale}
\item{FriendlAEF10s}{Item for the friendliness scale}
\item{FriendlAEF10u}{Item for the friendliness scale}
\item{FriendlAEK04x}{Item for the friendliness scale}
\item{FriendlAEK3x}{Item for the friendliness scale}
\item{FriendlAEK7x}{Item for the friendliness scale}
\item{HappyAEH01y}{Item for the happiness scale}
\item{HappyAEH03y}{Item for the happiness scale}
\item{HappyAEH04x}{Item for the happiness scale}
\item{HappyAEH06v}{Item for the happiness scale}
\item{HappyAEH06x}{Item for the happiness scale}
\item{HappyAEH1t}{Item for the happiness scale}
\item{InnerLiAOF06v}{Item for the inner life scale}
\item{InnerLiAOF07v}{Item for the inner life scale}
\item{InnerLiAOI01x}{Item for the inner life scale}
\item{InnerLiAOI04x}{Item for the inner life scale}
\item{InnerLiAOI06z}{Item for the inner life scale}
\item{InnerLiAOI2s}{Item for the inner life scale}
\item{IntelleAOIN01y}{Item for the intellectual scale}
\item{IntelleAOIN1x}{Item for the intellectual scale}
\item{IntelleAOIN3x}{Item for the intellectual scale}
\item{IntelleAOIN6x}{Item for the intellectual scale}
\item{IntelleAOIN8x}{Item for the intellectual scale}
\item{IntelleAOK10s}{Item for the intellectual scale}
\item{MoralAAM01J}{Item for the moralic scale}
\item{MoralAAM01x}{Item for the moralic scale}
\item{MoralAAM02v}{Item for the moralic scale}
\item{MoralAAM03s}{Item for the moralic scale}
\item{MoralAAM3m}{Item for the moralic scale}
\item{MoralAAM3p}{Item for the moralic scale}
\item{NonAnxiANA02y}{Item for the non anxiety scale}
\item{NonAnxiANA04y}{Item for the non anxiety scale}
\item{NonAnxiANA2r}{Item for the non anxiety scale}
\item{NonAnxiANA5s}{Item for the non anxiety scale}
\item{NonAnxiANA5u}{Item for the non anxiety scale}
\item{NonAnxiANA9s}{Item for the non anxiety scale}
\item{NonDeprAND1P}{Item for the non depressive scale}
\item{NonDeprAND3u}{Item for the non depressive scale}
\item{NonDeprAND4s}{Item for the non depressive scale}
\item{NonDeprAND9s}{Item for the non depressive scale}
\item{NonDeprANN12r}{Item for the non depressive scale}
\item{NonDeprANN7x}{Item for the non depressive scale}
\item{NonIrriANI10s}{Item for the non irritability scale}
\item{NonIrriANI1J}{Item for the non irritability scale}
\item{NonIrriANI7x}{Item for the non irritability scale}
\item{NonIrriANI8s}{Item for the non irritability scale}
\item{NonIrriANNI6x}{Item for the non irritability scale}
\item{NonIrriANNI8x}{Item for the non irritability scale}
\item{NonVulnANS11u}{Item for the non vulnerability scale}
\item{NonVulnANS2S}{Item for the non vulnerability scale}
\item{NonVulnANS5P}{Item for the non vulnerability scale}
\item{NonVulnANS8x}{Item for the non vulnerability scale}
\item{NonVulnANT6x}{Item for the non vulnerability scale}
\item{NonVulnANV01y}{Item for the non vulnerability scale}
\item{OrderlyACO03x}{Item for the orderly scale}
\item{OrderlyACO06x}{Item for the orderly scale}
\item{OrderlyACO12x}{Item for the orderly scale}
\item{OrderlyACO13x}{Item for the orderly scale}
\item{OrderlyACO1s}{Item for the orderly scale}
\item{OrderlyACO4c}{Item for the orderly scale}
\item{SelfEffACE01y}{Item for the self efficacy scale}
\item{SelfEffACE04y}{Item for the self efficacy scale}
\item{SelfEffACE05y}{Item for the self efficacy scale}
\item{SelfEffACE06y}{Item for the self efficacy scale}
\item{SelfEffACE14x}{Item for the self efficacy scale}
\item{SelfEffACE4J}{Item for the self efficacy scale}
\item{SympathAAS01y}{Item for the sympathy scale}
\item{SympathAAS02v}{Item for the sympathy scale}
\item{SympathAAS03v}{Item for the sympathy scale}
\item{SympathAAS9p}{Item for the sympathy scale}
\item{SympathAASY14}{Item for the sympathy scale}
\item{SympathAASY2e}{Item for the sympathy scale}
\item{VariatiAOV02v}{Item for the Variability seeking scale}
\item{VariatiAOV04v}{Item for the Variability seeking scale}
\item{VariatiAOV04x}{Item for the Variability seeking scale}
\item{VariatiAOV16x}{Item for the Variability seeking scale}
\item{VariatiAOV19s}{Item for the Variability seeking scale}
\item{VariatiAOV6x}{Item for the Variability seeking scale}
...
}
}
\source{
None
}
\usage{
persData
}
\description{
A dataset with 120 items, 20 scales and 5 factors (BigFive)
}
\keyword{datasets}
