% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot.Psychometric}
\alias{plot.Psychometric}
\title{Plot scales rom Psychometric}
\usage{
\method{plot}{Psychometric}(x, ...)
}
\arguments{
\item{x}{a Psychometric object}

\item{...}{extra argument is: scale = scaleName}
}
\value{
nothing
}
\description{
Plot scales rom Psychometric
}
\examples{
object <- GetPsychometric(persData, c("Achievement", "Dutifulness", "Orderly"),
 responseScale = list(c(0,4)), itemLength = 4)
 plot(object, scale = "Achievement")
}
