% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFacets.R
\name{print.TestFacets}
\alias{print.TestFacets}
\title{Title}
\usage{
\method{print}{TestFacets}(x, ...)
}
\arguments{
\item{x}{an object of class TestFacets}

\item{...}{model and standardized and more which model to print amd whether standardized solution should be printed}
}
\value{
summary of the lavaan results in the object
}
\description{
Title
}
