% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFacets.R
\name{RunCFA.Psychometric}
\alias{RunCFA.Psychometric}
\title{A Psychometric CFA estimator}
\usage{
\method{RunCFA}{Psychometric}(object, model, what = NULL, exclude = NULL)
}
\arguments{
\item{object}{an of class TestFacets}

\item{model}{a text string with an cfa model}
}
\description{
A Psychometric CFA estimator
}
