% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFacets.R
\name{RunCFA}
\alias{RunCFA}
\title{A Psychometric CFA estimator}
\usage{
RunCFA(object, model, what = NULL, exclude = NULL)
}
\arguments{
\item{object}{an of class TestFacets}

\item{model}{a text string with an cfa model}
}
\value{
a lavaan object with cfa results
}
\description{
A Psychometric CFA estimator
}
