% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Split.R
\name{split.Psychometric}
\alias{split.Psychometric}
\title{Split Psychometric}
\usage{
\method{split}{Psychometric}(x, f, drop, ...)
}
\arguments{
\item{x}{A Psychometric object}

\item{f}{A function to apply to the groups}

\item{drop}{A group variables among other variables}

\item{...}{group to split for and more arguments to the f function}
}
\value{
all results from the f function
}
\description{
Makes it simple to work with groups
}
