% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Split.R
\name{split.Psychometric}
\alias{split.Psychometric}
\title{split Psychometric}
\usage{
\method{split}{Psychometric}(x, f, drop = FALSE, ...)
}
\arguments{
\item{f}{A function to apply to the groups}

\item{drop}{logical indicating if levels that do not occur should be dropped (if f is a factor or a list).}

\item{...}{more arguments to the f function}

\item{object}{A Psychometric object}
}
\value{
all results from the f function
}
\description{
Makes it simple to work with groups
}
