% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPsychometric.R
\name{summary.Psychometric}
\alias{summary.Psychometric}
\title{summary for Psychometric}
\usage{
\method{summary}{Psychometric}(object, ...)
}
\arguments{
\item{object}{Psychometric object}

\item{...}{There are several extra arguments to delete from the output
mean: should the mean be displayed T/F
sd:  should the sd be displayed T/F
SE:  should the SE be displayed T/F
skew:  should the skew be displayed T/F
kurtosis:  should the kurtosis be displayed T/F
min:  should the min be displayed T/F
max:  should the max be displayed T/F
alpha:  should the alpha be displayed T/F. Warnings at Level
n:  should the n be displayed T/F
plots:  should the plots be displayed T/F}
}
\value{
Nothing
}
\description{
summary for Psychometric
}
\examples{
object <- GetPsychometric(persData, c("Achievement", "Dutifulness", "Orderly"),
   responseScale = list(c(0,4)), itemLength = 4)
summary(object)
}
