% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.R
\name{summary.Psychometric}
\alias{summary.Psychometric}
\title{Summary}
\usage{
\method{summary}{Psychometric}(object, ...)
}
\arguments{
\item{object}{The Psychometric object}

\item{...}{extra argument}
}
\description{
Includes a set of useful estimates for the scales
}
\examples{
object <- GetPsychometric(persData, c("Achievement", "Dutifulness", "Orderly"),
 responseScale = list(c(0,4)), itemLength = 4)
summary(object)
}
