% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.R
\name{summary.Psychometric}
\alias{summary.Psychometric}
\title{summary
Includes a set of useful estimates for the scales}
\usage{
\method{summary}{Psychometric}(object, ...)
}
\arguments{
\item{object}{The Psychometric object}

\item{...}{mean T: Shown
F: not shown
sd  T: Shown, F: not shown
SE  T: Shown, F: not shown
skew  T: Shown, F: not shown
kurtosis  T: Shown, F: not shown
min  T: Shown, F: not shown
max  T: Shown, F: not shown
omega  T: Shown, F: not shown
n  T: Shown, F: not shown
plots F: no plots are shown, T: plots from psych::omega are shown}
}
\value{
the summary i a list object
}
\description{
summary
Includes a set of useful estimates for the scales
}
\examples{
object <- GetPsychometric(persData, c("Achievement", "Dutifulness", "Orderly"),
 responseScale = list(c(0,4)), itemLength = 4)
summary(object)
}
