% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychometricHelper.R
\name{writeP}
\alias{writeP}
\title{writeP}
\usage{
writeP(object, File, colnames = T, rownames = F)
}
\arguments{
\item{object}{A psychometric object}

\item{colnames}{T if you like to write column names}

\item{rownames}{T if you like to write row names}

\item{fileName}{A filename with pathinformation}
}
\description{
writeP
}
